+++ src/engine/inject/additionalFunctions.js
@@ -1,3 +1,5 @@
+var no = require('not-defined')
+
 /**
  * Optionally add custom functions.
  *
@@ -7,9 +9,7 @@
 
 function injectAdditionalFunctions (funcs, additionalFunctions) {
   // Nothing to do if no additional function is given.
-  if (typeof additionalFunctions === 'undefined') {
-    return
-  }
+  if (no(additionalFunctions)) return
 
   /**
    * Validate and insert an additional function.
@@ -18,9 +18,7 @@ function injectAdditionalFunctions (funcs, additionalFunctions) {
   function injectAdditionalFunction (key) {
     var isAFunction = typeof additionalFunctions[key] === 'function'
 
-    if (isAFunction) {
-      funcs[key] = additionalFunctions[key]
-    }
+    if (isAFunction) funcs[key] = additionalFunctions[key]
   }
 
   Object.keys(additionalFunctions)